DHFPROD-4519: Move job-info under /models
