+++ one-ui/src/test/java/com/marklogic/hub/oneui/managers/ModelManagerTest.java
@@ -20,23 +20,48 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
+import com.marklogic.client.impl.DatabaseClientImpl;
+import com.marklogic.hub.HubConfig;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.platform.runner.JUnitPlatform;
-import org.junit.runner.RunWith;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.test.util.ReflectionTestUtils;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.when;
 
-@RunWith(JUnitPlatform.class)
 @ExtendWith(MockitoExtension.class)
 public class ModelManagerTest {
     @Mock
+    DatabaseClientImpl dataServiceClient;
+
+    @Mock
+    HubConfig hubConfig;
+
+    @Mock
     protected ModelManager modelManager;
+
+    private static ObjectMapper mapper;
+
+    @BeforeAll
+    static void setup() {
+        mapper = new ObjectMapper();
+    }
+
+    @AfterEach
+    void cleanup() {
+        ReflectionTestUtils.setField(modelManager, "finalDataServiceClient", null);
+    }
 
     private static final String PATIENT_ENTITY_MODEL = "[\n"
         + "  {\n"
@@ -321,23 +346,20 @@ public class ModelManagerTest {
 
     @Test
     public void testGetModels() throws IOException {
-        ObjectMapper mapper = new ObjectMapper();
         JsonNode expectObj = mapper.readTree(PATIENT_ENTITY_MODEL);
         when(modelManager.getModels()).thenReturn(expectObj);
         JsonNode resultObj = modelManager.getModels();
-        assertTrue(expectObj.equals(resultObj));
+        assertEquals(expectObj, resultObj);
     }
 
     @Test
     public void testGetModelByTitle() throws IOException {
-        ObjectMapper mapper = new ObjectMapper();
         JsonNode jsonNode = mapper.readTree(PATIENT_ENTITY_MODEL);
         ArrayNode entity = JsonNodeFactory.instance.arrayNode();
         if (jsonNode.isArray()) {
             jsonNode.forEach(e -> {
                 if ("Labs".equals(e.get("info").get("title").asText())) {
                     entity.add(e.toString());
-                    return;
                 }
             });
         }
@@ -348,19 +370,111 @@ public class ModelManagerTest {
 
     @Test
     public void testGetModelNoSuchTitle() throws IOException {
-        ObjectMapper mapper = new ObjectMapper();
         JsonNode jsonNode = mapper.readTree(PATIENT_ENTITY_MODEL);
         ArrayNode entity = JsonNodeFactory.instance.arrayNode();
         if (jsonNode.isArray()) {
             jsonNode.forEach(e -> {
                 if ("Lab".equals(e.get("info").get("title").asText())) {
                     entity.add(e.toString());
-                    return;
                 }
             });
         }
         when(modelManager.getModel("Lab")).thenReturn(entity);
         JsonNode resultObj = modelManager.getModel("Lab");
         assertTrue(resultObj.size() == 0 && entity.equals(resultObj));
+    }
+
+    @Test
+    void testLatestJobInfoForModel() throws IOException {
+        // setup
+        String resp = "{\n"
+            + "        \"modelName\": \"Product\",\n"
+            + "        \"jobId\": \"ab6cb392-2f9c-40ce-bd83-b15a62b6aa84\",\n"
+            + "        \"jobTime\": \"2019-09-27T13:08:11.467812-07:00\"\n"
+            + "    }";
+        JsonNode expectedNode = mapper.readTree(resp);
+        modelManager = Mockito.spy(new ModelManager(hubConfig));
+        ReflectionTestUtils.setField(modelManager, "finalDataServiceClient", dataServiceClient);
+        String modelName = "Product";
+        doReturn(expectedNode).when(modelManager).getJobInfoFromDB(dataServiceClient, modelName);
+
+        // test
+        JsonNode actualNode = modelManager.getLatestJobInfo(modelName);
+
+        // verify
+        assertEquals(expectedNode, actualNode);
+    }
+
+    @Test
+    void testLatestJobInfoForNonExistentModel() throws IOException {
+        // setup
+        String resp = "{\n"
+            + "        \"modelName\": \"xyz\",\n"
+            + "        \"jobId\": null,\n"
+            + "        \"jobTime\": null\n"
+            + "    }";
+        JsonNode expectedNode = mapper.readTree(resp);
+        modelManager = Mockito.spy(new ModelManager(hubConfig));
+        ReflectionTestUtils.setField(modelManager, "finalDataServiceClient", dataServiceClient);
+        String modelName = "xyz";
+        doReturn(expectedNode).when(modelManager).getJobInfoFromDB(dataServiceClient, modelName);
+
+        // test
+        JsonNode actualNode = modelManager.getLatestJobInfo(modelName);
+
+        // verify
+        assertEquals(expectedNode, actualNode);
+    }
+
+    @Test
+    void testLatestJobInfoForAllModels() throws IOException {
+        // setup
+        String resp = "[\n"
+            + "    {\n"
+            + "        \"modelName\": \"Order\",\n"
+            + "        \"jobId\": null,\n"
+            + "        \"jobTime\": null\n"
+            + "    },\n"
+            + "    {\n"
+            + "        \"modelName\": \"Product\",\n"
+            + "        \"jobId\": \"ab6cb392-2f9c-40ce-bd83-b15a62b6aa84\",\n"
+            + "        \"jobTime\": \"2019-09-27T13:08:11.467812-07:00\"\n"
+            + "    }\n"
+            + "]";
+        JsonNode expectedNode = mapper.readTree(resp);
+
+        String product = "{\n"
+            + "        \"modelName\": \"Product\",\n"
+            + "        \"jobId\": \"ab6cb392-2f9c-40ce-bd83-b15a62b6aa84\",\n"
+            + "        \"jobTime\": \"2019-09-27T13:08:11.467812-07:00\"\n"
+            + "    }";
+        JsonNode productNode = mapper.readTree(product);
+        modelManager = Mockito.spy(new ModelManager(hubConfig));
+        doReturn(productNode).when(modelManager)
+            .getJobInfoFromDB(dataServiceClient, "Product");
+
+        String order = "{\n"
+            + "        \"modelName\": \"Order\",\n"
+            + "        \"jobId\": null,\n"
+            + "        \"jobTime\": null\n"
+            + "    }";
+        JsonNode orderNode = mapper.readTree(order);
+        doReturn(orderNode).when(modelManager)
+            .getJobInfoFromDB(dataServiceClient, "Order");
+
+        ReflectionTestUtils.setField(modelManager, "finalDataServiceClient", dataServiceClient);
+
+        List<String> modelNameList = new ArrayList<>();
+        modelNameList.add("Order");
+        modelNameList.add("Product");
+        doReturn(modelNameList).when(modelManager).getModelNames();
+
+        // test
+        List<JsonNode> actualList = modelManager.getLatestJobInfoForAllModels();
+        String actualString = mapper.writeValueAsString(actualList);
+        JsonNode actualNode = mapper.readTree(actualString);
+
+        // verify
+        assertEquals(expectedNode, actualNode);
     }
 }
